+++ src/Imi.Project.Api.Core/Interfaces/Repository/IFavoriteRepository.cs
@@ -6,8 +6,8 @@ namespace Imi.Project.Api.Core.Interfaces.Repository
 {
     public interface IFavoriteRepository : IRepository<Favorite>
     {
-        Task<Favorite> GetByUserIdAndMovieIdAsync(string userId, long movieId);
-        Task<Favorite> AddByUserIdAndMovieIdAsync(string userId, long movieId);
+        Task<Favorite> GetByUserIdAndMovieIdAsync(string userId, int movieId);
+        Task<Favorite> AddByUserIdAndMovieIdAsync(string userId, int movieId);
         Task<IEnumerable<Favorite>> GetFavoritesByUserIdAsync(string userId);
     }
 }
