+++ src/Imi.Project.Api.Core/Interfaces/Service/IService.cs
@@ -7,10 +7,10 @@ namespace Imi.Project.Api.Core.Interfaces.Service
     public interface IService<TResponse, Trequest> where TResponse : BaseDto where Trequest : BaseDto
     {
         Task<IEnumerable<TResponse>> ListAllAsync();
-        Task<TResponse> GetByIdAsync(long id);
+        Task<TResponse> GetByIdAsync(int id);
         Task<TResponse> AddAsync(Trequest RequestDto);
         Task<TResponse> UpdateAsync(Trequest RequestDto);
-        Task DeleteAsync(long id);
+        Task DeleteAsync(int id);
 
     }
 }
