+++ src/Imi.Project.Common/Dtos/FavoriteResponseDto.cs
@@ -4,7 +4,7 @@ namespace Imi.Project.Common.Dtos
 {
     public class FavoriteResponseDto : BaseDto
     {
-        public long MovieId { get; set; }
+        public int MovieId { get; set; }
         public string ApplicationUserId { get; set; }
         public MovieResponseDto Movie { get; set; }
         public UserResponseDto User { get; set; }
