+++ src/Imi.Project.Wpf.Core/Entities/MovieGenre.cs
@@ -4,10 +4,10 @@ namespace Imi.Project.Wpf.Core.Entities
 {
     public class MovieGenre : BaseEntity
     {
-        public long MovieId { get; set; }
+        public int MovieId { get; set; }
         public Movie Movie { get; set; }
 
-        public long GenreId { get; set; }
+        public int GenreId { get; set; }
         public Genre Genre { get; set; }
 
     }
